fix: add new homepage
